requirements.txt
# Python version
python==3.9

# Project dependencies
torch==2.0.1
torchvision==0.15.2
matplotlib==3.7.1
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
scikit-learn==1.2.2
Pillow==9.5.0
opencv-python-headless==4.7.0.72
tqdm==4.65.0
seaborn==0.12.2